requirements.txt
yfinance
pandas
numpy
schedule
requests
python-telegram-bot[job-queue]
streamlit
plotly
python-dotenv